+++ module/auth-rest/e2e/src/conf.ts
@@ -3,6 +3,7 @@ import { InjectableFactory } from '@travetto/di';
 import { MemoryStore, SessionStore, SessionEncoder, CookieEncoder } from '@travetto/rest-session';
 
 import { IdentityProvider } from '../..';
+import { AuthContextStore, SessionAuthContextStore } from '../../src/store';
 
 export class FbUser {
   id: string;
@@ -21,6 +22,11 @@ export class AppConfig {
   @InjectableFactory()
   static sessionEncoder(): SessionEncoder {
     return new CookieEncoder();
+  }
+
+  @InjectableFactory()
+  static sessionAuthStore(): AuthContextStore {
+    return new SessionAuthContextStore();
   }
 
   @InjectableFactory()
