+++ module/auth-rest/index.ts
@@ -4,5 +4,5 @@ export * from './src/decorator';
 export * from './src/identity';
 export * from './src/errors';
 export * from './src/service';
-export * from './src/state';
+export * from './src/store';
 export * from './src/adapter';
